Move common utility code to plugins.module_utils.util.
