+++ plugins/modules/docker_image_load.py
@@ -77,9 +77,11 @@ from ansible.module_utils.common.text.converters import to_native
 
 from ansible_collections.community.docker.plugins.module_utils.common import (
     AnsibleDockerClient,
+    RequestException,
+)
+from ansible_collections.community.docker.plugins.module_utils.util import (
     DockerBaseClass,
     is_image_name_id,
-    RequestException,
 )
 
 try:
