+++ plugins/modules/docker_plugin.py
@@ -129,10 +129,12 @@ except ImportError:
     pass
 
 from ansible_collections.community.docker.plugins.module_utils.common import (
-    DockerBaseClass,
     AnsibleDockerClient,
-    DifferenceTracker,
     RequestException
+)
+from ansible_collections.community.docker.plugins.module_utils.util import (
+    DockerBaseClass,
+    DifferenceTracker,
 )
